+++ R/two-samp-var-test-shiny.R
@@ -1,11 +1,11 @@
 #' @importFrom stats complete.cases
 #' @importFrom purrr map_dbl
-#' @title Two Sample Variance Comparison Test
+#' @title Two Sample Variance Comparison Test Shiny
 #' @description  \code{var_test} performs tests on the equality of standard
 #' deviations (variances).
-#' @param variable a numeric vector
-#' @param group_var a grouping variable
-#' @param ... numeric vectors
+#' @param data a dataframe
+#' @param variable1 character; name of sample 1 in data frame
+#' @param variable2 character; name of sample 2 in data frame
 #' @param alternative a character string specifying the alternative hypothesis,
 #' must be one of "both" (default), "greater", "less" or "all". You can specify
 #' just the initial letter.
@@ -35,67 +35,29 @@
 #' Statistical Procedures, 4th edition. : Chapman & Hall/CRC.
 #' @seealso \code{\link[stats]{var.test}}
 #' @examples
-#' # using grouping variable
-#' # lower tail
-#' var_test(mtcars$mpg, group_var = mtcars$vs, alternative = 'less')
-#'
-#' # upper tail
-#' var_test(mtcars$mpg, group_var = mtcars$vs, alternative = 'greater')
-#'
-#' # all tails
-#' var_test(mtcars$mpg, group_var = mtcars$vs, alternative = 'all')
-#'
 #' # using two variables
 #' # lower tail
-#' var_test(hsb$read, hsb$write, alternative = 'less')
+#' var_test_shiny(hsb, 'read', 'write', alternative = 'less')
 #'
 #' # upper tail
-#' var_test(hsb$read, hsb$write, alternative = 'greater')
+#' var_test_shiny(hsb, 'read', 'write', alternative = 'greater')
 #'
 #' # all tails
-#' var_test(hsb$read, hsb$write, alternative = 'all')
+#' var_test_shiny(hsb, 'read', 'write', alternative = 'all')
 #'
 #' @export
 #'
-var_test <- function(variable, ..., group_var = NA,
-	alternative = c("less", "greater", "all")) UseMethod('var_test')
+var_test_shiny <- function(data, variable1, variable2, 
+	alternative = c("less", "greater", "all")) UseMethod('var_test_shiny')
 
 #' @export
 #'
-var_test.default <- function(variable, ..., group_var = NA,
+var_test_shiny.default <- function(data, variable1, variable2, 
 	alternative = c("less", "greater", "all")) {
 
-	if (length(group_var) != 1) {
-		if (nlevels(as.factor(group_var)) != 2) {
-			stop('group_var must be a binary factor variable.', call. = FALSE)
-		}
-	}
-
-	suppressWarnings(
-		if (is.na(group_var)) {
-			name1 <- l(unlist(strsplit(deparse(substitute(c(variable, ...))), ','))[1])
-			name2 <- unlist(strsplit(l(unlist(strsplit(deparse(substitute(c(variable, ...))), ','))[2]), ')'))[1]
-			lev <- c(name1, name2)
-		} else {
-			if (!is.factor(group_var)) {
-				group_var <- as.factor(group_var)
-			}
-			if (nlevels(group_var) > 2) {
-				stop('Specify a binary factor variable as input for group_var.', call. = FALSE)
-			}
-			lev  <- levels(group_var)
-		}
-	)
-
-	suppressWarnings(
-
-		if (is.na(group_var)) {
-
-			z   <- list(variable, ...)
+			variables <- data %>% select_(variable1, variable2)
+			z   <- list(variables[[1]], variables[[2]])
 			ln <- z %>% map_int(length)
-			if (length(ln) > 2) {
-				stop('Only 2 variables can be specified.', call. = FALSE)
-			}
 			ly <- seq_len(length(z))
 
 		  if (length(z) < 2) {
@@ -106,31 +68,22 @@ var_test.default <- function(variable, ..., group_var = NA,
 		  variable  <- unlist(z)
 		  group_var <- unlist(out)
 
-		} else {
-
-    	if (length(variable) != length(group_var)) {
-    		stop('Length of variable and group_var do not match.', call. = FALSE)
-    	}
-
-    }
-
-	)
-
 	type <- match.arg(alternative)
 	   k <- var_comp(variable, group_var)
+	   lev <- c(variable1, variable2)
 	   
 	out <- list(f = k$f, lower = k$lower, upper = k$upper, vars = k$vars,
          avgs = k$avgs, sds = k$sds, ses = k$ses, avg = k$avg, sd = k$sd,
          se = k$se, n1 = k$n1, n2 = k$n2, lens = k$lens, len = k$len,
          lev = lev, type = type)
 
-	class(out) <- 'var_test'
+	class(out) <- 'var_test_shiny'
   return(out)
 
 }
 
 #' @export
 #'
-print.var_test <- function(x, ...) {
+print.var_test_shiny <- function(x, ...) {
   print_var_test(x)
 }